XML scene reader finished!!
The .xml geometry parser is done! Now default options can be
established. That is: instead of having to write 'bord="1"' for every
obstacle with border 1, we can write the attribute 'bord="1"' in the
node 'obstacles' (or 'enemy', or 'walls') and every obstacle that
doesn't specify a border width will have border 1!