"Spring" back home effect added
When the player leaves "home", it is forced to come back after a short
time. Even if the player keeps pressing the key to move away from home,
the "force" of the spring gets stronger and stronger until it reaches
the home line again. (play the game for a better description lol)