Main menu added!
When the game is started, a main menu is displayed with "n" levels
shown (all locked except for the first one) and an "Exit" option. When
user selects a level, it is loaded from the file "Level X.xml" (where X
is the level number). The menu is displayed again when the player
either dies or completes the level.